fixed tests and updated conftest fixtures
